requirements.txt
httpx>=0.27.0
fastapi>=0.110.0
uvicorn>=0.29.0
pystray>=0.19.5
Pillow>=10.0.0
